requirements.txt
Bio-Epidemiology-NER==0.1.3
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
fsspec==2023.10.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
openai-whisper
packaging==23.2
pandas==1.4.2
Pillow==10.1.0
PyAudio==0.2.13
PyMuPDF==1.20.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six==1.16.0
sympy==1.12
tiktoken==0.3.3
tokenizers==0.14.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.0
zipp==3.17.0
